proxy: move embedded envoy aspects into dedicated file
This commit refactors all aspects of the embedded Envoy proxy instance
into a separate file `embedded_envoy.go` and separates it from other
aspects like EnvoyAdminClient.
